+++ lmctl/cli/commands/pkg.py
@@ -1,5 +1,7 @@
 import click
 import logging
+import shutil
+import os
 import lmctl.cli.lifecycle as lifecycle_cli
 import lmctl.project.package.core as pkgs
 from lmctl.cli.format import determine_format_class
@@ -25,10 +27,13 @@ def push(package, environment, config, armname, pwd):
     """Pushes an existing Assembly/Resource package to a target LM (and ARM) environment"""
     logger.debug('Pushing package at: {0}'.format(package))
     pkg = lifecycle_cli.open_pkg(package)
-    env_sessions = lifecycle_cli.build_sessions_for_project(package.config, environment, pwd, armname, config)
-    controller = lifecycle_cli.ExecutionController(PUSH_HEADER)
-    controller.start(pkg.path)
-    exec_push(controller, pkg, env_sessions)
+    try:
+        env_sessions = lifecycle_cli.build_sessions_for_pkg(pkg.meta, environment, pwd, armname, config)
+        controller = lifecycle_cli.ExecutionController(PUSH_HEADER)
+        controller.start(package)
+        exec_push(controller, pkg, env_sessions)
+    finally:
+        cleanup_pkg(pkg)
     controller.finalise()
 
 
@@ -39,10 +44,17 @@ def push(package, environment, config, armname, pwd):
 def inspect(package, config, output_format):
     logger.debug('Inspecting package at: {0}'.format(package))
     pkg = lifecycle_cli.open_pkg(package)
-    inspection_report = pkg.inspect()
-    result = format_inspection_report(output_format, inspection_report)
-    click.echo(result)
+    try:
+        inspection_report = pkg.inspect()
+        result = format_inspection_report(output_format, inspection_report)
+        click.echo(result)
+    finally:
+        cleanup_pkg(pkg)
     
+def cleanup_pkg(pkg):
+    if os.path.exists(pkg.tree.root_path):
+        shutil.rmtree(pkg.tree.root_path)
+
 def format_inspection_report(output_format, inspection_report):
     inspection_report_tpl = inspection_report.to_dict()
     formatter_class = determine_format_class(output_format)
@@ -53,4 +65,4 @@ def format_inspection_report(output_format, inspection_report):
 def exec_push(controller, pkg, env_sessions):
     push_options = pkgs.PushOptions()
     push_options.journal_consumer = controller.consumer
-    return controller.execute(pkg.push, env_sessions)
+    return controller.execute(pkg.push, env_sessions, push_options)