+++ lmctl/utils/certificates.py
@@ -1,0 +1,5 @@
+def read_certificate_file(path):
+    with open(path, 'rt') as f:
+        certificate = f.read()
+        certificate = certificate.replace("\n", "\\n")
+        return certificate