+++ tests/integration/project/assemblytests/test_create.py
@@ -35,7 +35,8 @@ class TestCreateAssemblyProjects(ProjectSimTestCase):
             'schema':  '2.0',
             'name': 'Test',
             'version': '1.0',
-            'type': 'Assembly'
+            'type': 'Assembly',
+            'packaging': 'tgz'
         })
         
     def test_create(self):
@@ -56,7 +57,8 @@ class TestCreateAssemblyProjects(ProjectSimTestCase):
             'schema':  '2.0',
             'name': 'Test',
             'version': '9.9',
-            'type': 'Assembly'
+            'type': 'Assembly',
+            'packaging': 'tgz'
         })
 
     def test_create_with_subprojects(self):
@@ -85,6 +87,7 @@ class TestCreateAssemblyProjects(ProjectSimTestCase):
             'schema':  '2.0',
             'name': 'Test',
             'version': '9.9',
+            'packaging': 'tgz',
             'type': 'Assembly',
             'contains': [
                 {